Rename handleMovingCursor to adjustCursorIfAtEdge
Also clean up a bit.
